Finished execute stage
Execute stage is finished.  Just need an additional 2-bit BranchOp
signal from ALU that will choose the appropriate zero equality to use
for the branches (just use 00 if no branch occurring).  Also need to
add invSrc1 and invSrc2 signals from control to tell ALU if either of
the source signals to the ALU need to be inverted, and also a sub
signal to let the ALU know if subtraction is occurring.